Update colours on link options toggle
When "Calculate colors" or "Title bar gradients" is toggled, update the
display right away. The state is committed to history and restored on
undo.

Previously, the user would have to trigger a change on ButtonFace or
ActiveTitle/InactiveTitle for the change to be visible, which is not
intuitive.

TODO: Toggling "Title bar gradients" should save and restore the
Gradient colours if the colour has not been changed. This was not done
at this time since the user can undo to get back to the previous state,
but it is not intuitive as the user would expect to see the gradient
restored when re-enabling the option.